Correct P010, P012, P016 rendering
Switch the inherited class for the derived ImageRenderer classes for these
image types to `BaseYuv420sp16ImageRenderer` from
`BaseYuv420p16ImageRenderer`; it looks like this was was the intended
class to inherit from due to these classes being defined in the
BaseYuv420sp16ImageRenderer source file.

As well, change the constructor for these to only initialize two planes
instead of three since the P01* image types are semi-planar formats,
similar to NV12.

Finally, change the effective bits for all of these types to 16 bits;
the idea behind the format is that for P010 and P012 the unused lower bits are
0 and so they can be interpreted as less-detailed versions of P016.

Signed-off-by: Eric Joyner <erj@erj.cc>